Fix formatting, separate song context hooks file
