+++ frontend/lyric-sync-app/src/api/apis/index.ts
@@ -1,5 +1,5 @@
 /* tslint:disable */
-/* eslint-disable */
+ 
 export * from './HandwritingApi';
 export * from './SongApi';
 export * from './UserApi';
