+++ frontend/lyric-sync-app/src/api/index.ts
@@ -1,5 +1,5 @@
 /* tslint:disable */
-/* eslint-disable */
+ 
 export * from './runtime';
 export * from './apis/index';
 export * from './models/index';
