+++ frontend/lyric-sync-app/src/api/models/index.ts
@@ -1,5 +1,5 @@
 /* tslint:disable */
-/* eslint-disable */
+ 
 export * from './Album';
 export * from './FullSongRequestDto';
 export * from './FullSongResponseDto';
